all employer and jobs controllers

+++ app/Http/Controllers/EmployerController.php
@@ -4,54 +4,25 @@ namespace App\Http\Controllers;
 
 use App\Models\Employer;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class EmployerController extends Controller
 {
-
     function index()
     {
         $employer = Employer::all();
         return response()->json($employer,200);
 
     }
-    function store(Request $request)
+
+    public function update(Request $request)
     {
-        $validated = $request->validate([
-            'user_id' => 'required|integer',
-            'company_name' => 'required|string|max:255',
-            'description' => 'nullable|string',
-            'industry' => 'nullable|string',
-            'company_size' => 'nullable|string',
-            'established_date' => 'nullable|date',
-            'logo_url' => 'nullable|image|mimes:jpg,jpeg,png,svg|max:2048',
-        ]);
+        $user = Auth::user();
+        $employer = Employer::where('user_id', $user->user_id)->first();
 
-
-        if ($request->hasFile('logo_url')) {
-            $validated['logo_url'] = $request->file('logo_url')->store('employersLogos', 'public');
+        if (!$employer) {
+            return response()->json(['message' => 'Employer not found'], 404);
         }
-
-        $employer = Employer::create($validated);
-
-        return response()->json([
-            'message' => "Employer created",
-            'employer' => $employer,
-        ], 201);
-
-    }
-
-//    function update(Request $request,$id = 1)
-//    {
-//        $employer = Employer::findOrFail($id);
-//        $employer->update(
-//            $request->only('company_name', 'description', 'industry', 'logo_url', 'company_size', 'established_date')
-//        );
-//
-//    }
-
-    public function update(Request $request, $id = 1)
-    {
-        $employer = Employer::findOrFail(1);
 
         if ($request->hasFile('logo_url')) {
             $file = $request->file('logo_url');
@@ -71,15 +42,27 @@ class EmployerController extends Controller
     }
 
 
-    function show($id){
-        $employer = Employer::findOrFail($id);
-        return response()->json($employer,200);
+    function show(){
+        $user = Auth::user();
+        $employer = Employer::where('user_id', $user->user_id)->first();
+
+        if (!$employer) {
+            return response()->json(['message' => 'Employer not found'], 404);
+        }
+
+        return response()->json($employer, 200);
     }
 
-    function destroy($id){
-        $employer = Employer::findOrFail($id);
+    function destroy(){
+        $user = Auth::user();
+
+        $employer = Employer::where('user_id', $user->user_id);
+
+        if (!$employer) {
+            return response()->json(['message' => 'Employer not found'], 404);
+        }
+
         $employer->delete();
         return response()->json(null,204);
-
     }
 }
